[test] Address review feedback; add #15993 browser-baking coverage
PR feedback (#15977):
- Fold TestHeartbeatOTelMonitors into heartbeat_monitoring_test.go and drop
  the separate heartbeat_otel_monitoring_test.go, reusing the existing ess
  synthetics/heartbeat machinery (swiatekm).
- Deduplicate the OTel runtime override: switchHeartbeatToOtelRuntime now
  calls heartbeatOtelRuntimeOverride() (ycombinator).
- Bump TestOTelComponentsHealthy deadline 10m -> 20m to cover four serial
  5-minute Eventually checks (ycombinator).

Coverage for #15993 (Playwright browsers not baked into the complete image):
- package_test.go: assert the complete docker image's .cache/ms-playwright
  cache actually contains the Chromium (headless-shell + chrome) binaries.
- k8s standalone kustomize (complete image only): run a real inline browser
  journey and require it to launch Chromium and pass — a component-health
  check alone does not catch #15993.

Tests only; the product fix lives in #15994.
